SNOW-2043816: SEparate file
